# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users/me": {
      "get": {
        "operationId": "UsersController_getMe",
        "summary": "Get authenticated user profile",
        "tags": ["Users"],
        "security": [{ "bearer": [] }],
        "responses": {
          "200": { "description": "User profile" }
        }
      },
      "patch": {
        "operationId": "UsersController_updateMe",
        "summary": "Update authenticated user profile (fullName, avatarUrl)",
        "tags": ["Users"],
        "security": [{ "bearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateUserDto" }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated user profile" }
        }
      },
      "delete": {
        "operationId": "UsersController_deleteMe",
        "summary": "Permanently delete account and all data (GDPR Right to Erasure)",
        "tags": ["Users"],
        "security": [{ "bearer": [] }],
        "responses": {
          "200": { "description": "Account deleted. All personal data erased." },
          "404": { "description": "User not found" }
        }
      }
    },
    "/api/users/me/export": {
      "get": {
        "operationId": "UsersController_exportData",
        "summary": "Export all personal data as JSON (GDPR Right of Access)",
        "tags": ["Users"],
        "security": [{ "bearer": [] }],
        "responses": {
          "200": { "description": "JSON dump of all user data across all tenants" }
        }
      }
    },
...
    "/api/webhooks/mux": {
      "post": {
        "operationId": "WebhooksController_handleMuxWebhook",
        "summary": "Mux webhook receiver (HMAC-SHA256 verified, no JWT)",
        "tags": ["Webhooks"],
        "responses": {
          "200": { "description": "{ received: true }" },
          "401": { "description": "Invalid or missing mux-signature header" }
        }
      }
    },
...
  },
  "info": {
    "title": "ChurchApp Platform API",
    "description": "Multi-tenant church management platform & global social network.\n\n**Authentication:** All protected endpoints require a Supabase JWT in the `Authorization: Bearer <token>` header.\n\n**Tenant Context:** Most endpoints are tenant-scoped via PostgreSQL Row-Level Security. Call `POST /api/auth/switch-tenant` followed by `POST /api/auth/refresh` to set the active tenant in your JWT.\n\n**Rate Limits:** Global: 100 req/min per IP. Auth endpoints: 5 req/min per IP. Webhooks: unlimited.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    { "url": "http://localhost:3000", "description": "Local development" }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Supabase JWT access token"
      }
    },
    "schemas": {
...
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "fullName": { "type": "string", "maxLength": 255, "example": "John Doe" },
          "avatarUrl": { "type": "string", "format": "uri", "maxLength": 2048, "example": "https://cdn.example.com/avatar.jpg" }
        }
      },
...
    }
  }
}